Refactored api call I hoped to use to delete all the dinner cards on the calendar. Ran into a bug while testing where different users see the same calendar view

+++ pages/index.js
@@ -10,23 +10,24 @@ function Home() {
   const [days, setDays] = useState([]);
   const { user } = useAuth();
 
+  const deleteDinners = () => {
+    if (window.confirm('Are you sure you want to clear all dinners currently listed for this week? Click "OK" if you wish to continue.')) {
+      deleteAllDinnerCards(user.uid);
+    }
+  };
+
   const getSunThruSat = () => {
     getDaysOfTheWeek().then(setDays);
   };
 
-  const deleteAllDinners = () => {
-    deleteAllDinnerCards(user.uid);
-  };
-
   useEffect(() => {
     getSunThruSat();
-    deleteAllDinners();
   }, []);
 
   return (
     <div>
       <h2>Dinners This Week</h2>
-      <Button variant="secondary">Clear the Dinners for This Week</Button>
+      <Button variant="secondary" onClick={deleteDinners}>Clear the Dinners for This Week</Button>
       <div className="days-container">
         {days?.map((day) => (
           <DinnerCards key={day.firebaseKey} dayObj={day} />
